Refactor OfferEvaluator to use more general pipeline approach, add dynamic/named vip ports (#413)
This PR removes FulfilledRequirement and changes logic in OfferEvaluator to evaluate offers using a general queue based on a series of stages in a pipeline. Each stage takes a MesosResourcePool and OfferRequirement as input and generates an OfferRecommendation and/or modifies the TaskInfos and ExecutorInfo in the OfferRequirement as necessary. Stages for creating dynamic ports based on the available ports in the offer and for creating a DC/OS named VIP mapping to a port are included in this PR.